+++ src/lib/jp2/util/GrkImage.cpp
@@ -172,7 +172,7 @@ bool GrkImage::copyHeader(GrkImage *dest) {
 			pal_dest->component_mapping = new grk_component_mapping_comp[pal_dest->num_channels];
 			memcpy(pal_dest->component_mapping, pal_src->component_mapping,
 									pal_src->num_channels * sizeof(grk_component_mapping_comp));
-			memcpy(pal_dest->lut, pal_src->lut, pal_src->num_channels * pal_src->num_entries * sizeof(uint32_t));
+			memcpy(pal_dest->lut, pal_src->lut, (size_t)pal_src->num_channels * pal_src->num_entries * sizeof(uint32_t));
 		}
 	}
